core is fully deployed strategy
the strategy is a bit spread in different parts, might need to be refactored (I would very much like a strategy pattern to be centralized in one place, I call this strategy more because I want it to be one than because it currently is..), but the idea is the following:

there can be two ways to deploy the relationships:

- option A (fully deployed): the core contains all the input data. It is all joined based on their relationships and kept for further -right away- use

- option B (only keys and indexes deployed): the original dimension are kept on the original tables, not deployed on the core. On query execution, we join the dimensions that we need only to fetch them and bring them to the queries.

Trade off explained:

Basically if we fully deploy all the dimensions we don't need to do it on query execution, but the core grows in size with the data and the relationships. If we do it on the fly (option b) then the core is smaller but we need to join the dimensions to it as much as we need on the queries.

So a dashboard can have several plots which are using queries, the more queries the more we would need to join back to the core on option B.

For filtering, masking is used, so the core is basically copy/pasted with the filtered relationships. The bigger the core the longer this will take.

Performance of a vs b will depend on the data size, the relationships and the processing capacity. For small snowflake models fully deployed is working better. I haven't fully tested it yet but I assume that with bigger models memory will be a limitation for this strategy.